+++ scalingo/utils.go
@@ -31,7 +31,7 @@ func SetAll(d *schema.ResourceData, values map[string]interface{}) error {
 	for name, value := range values {
 		err := d.Set(name, value)
 		if err != nil {
-			return err
+			return fmt.Errorf("fail to set field %s: %v", name, err)
 		}
 	}
 	return nil