+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## To be Released
 
+* chore(errors) Switch error management to fmt.Errorf
 * feature(stacks) Add support for Scalingo stacks
 * chore(go-scalingo) Upgrade to go-scalingo v4
 * chore(terraform-plugin-sdk) Upgrade to SDK v2